+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Histograms are tracked for array and string lengths
 - JSON values can be checked for anomalies against a schema
+- JSON schemas can be converted to JSONoid objects
 
 ### Changed
 - Better names for generated definitions